Refactor `user.proto` files and update Go dependencies.
Deleted `user.proto` and refactored its corresponding Go files. Updated Go module dependencies and versioning to ensure compatibility and improved functionality.
