--- api/proto/user.proto
@@ -1,26 +1,0 @@
-syntax = "proto3";
-
-package user;
-
-import "google/protobuf/empty.proto";
-
-option go_package = "pkg/proto;usergen";
-
-service UserService {
-  rpc GetUser (GetUserRequest) returns (GetUserResponse);
-  rpc CreateUser (CreateUserRequest) returns (google.protobuf.Empty);
-}
-
-message GetUserRequest {
-  string username = 1;
-}
-
-message GetUserResponse {
-  string UUID = 1;
-}
-
-message CreateUserRequest {
-  string UUID = 1;
-  string username = 2;
-  string email = 3;
-}
